methods moved to functions.php
