+++ tests/TestCase/Core/FunctionsGlobalTest.php
@@ -20,7 +20,6 @@ use Cake\Core\Configure;
 use Cake\Http\Response;
 use Cake\TestSuite\TestCase;
 use stdClass;
-use function env;
 
 require_once CAKE . 'Core/functions_global.php';
 
@@ -337,5 +336,29 @@ class FunctionsGlobalTest extends TestCase
             triggerWarning('This was a mistake.');
             $this->assertTrue(true);
         });
+    }
+
+    /**
+     * @dataProvider \Cake\Test\TestCase\Core\FunctionsTest::toStringProvider
+     */
+    public function testToString(mixed $rawValue, ?string $expected): void
+    {
+        $this->assertSame($expected, toString($rawValue));
+    }
+
+    /**
+     * @dataProvider \Cake\Test\TestCase\Core\FunctionsTest::toIntProvider
+     */
+    public function testToInt(mixed $rawValue, null|int $expected): void
+    {
+        $this->assertSame($expected, toInt($rawValue));
+    }
+
+    /**
+     * @dataProvider \Cake\Test\TestCase\Core\FunctionsTest::toBoolProvider
+     */
+    public function testToBool(mixed $rawValue, ?bool $expected): void
+    {
+        $this->assertSame($expected, toBool($rawValue));
     }
 }
